+++ src/main/java/roomescape/reservation/application/ReservationFacadeImpl.java
@@ -3,8 +3,8 @@ package roomescape.reservation.application;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import roomescape.reservation.application.dto.AvailableReservationTimeServiceRequest;
-import roomescape.reservation.application.usecase.ReservationCommandService;
-import roomescape.reservation.application.usecase.ReservationQueryService;
+import roomescape.reservation.application.service.ReservationCommandService;
+import roomescape.reservation.application.service.ReservationQueryService;
 import roomescape.reservation.domain.ReservationDate;
 import roomescape.reservation.domain.ReservationId;
 import roomescape.reservation.ui.dto.AvailableReservationTimeWebResponse;
