+++ src/main/java/roomescape/reservation/application/service/ReservationQueryServiceImpl.java
@@ -1,4 +1,4 @@
-package roomescape.reservation.application.usecase;
+package roomescape.reservation.application.service;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -12,7 +12,7 @@ import roomescape.reservation.domain.Reservation;
 import roomescape.reservation.domain.ReservationDate;
 import roomescape.reservation.domain.ReservationRepository;
 import roomescape.theme.domain.ThemeId;
-import roomescape.time.application.usecase.ReservationTimeQueryService;
+import roomescape.time.application.service.ReservationTimeQueryService;
 import roomescape.time.domain.ReservationTime;
 import roomescape.time.domain.ReservationTimeId;
