+++ src/main/java/roomescape/theme/application/ThemeFacadeImpl.java
@@ -3,8 +3,8 @@ package roomescape.theme.application;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import roomescape.reservation.domain.ReservationDate;
-import roomescape.theme.application.usecase.ThemeCommandService;
-import roomescape.theme.application.usecase.ThemeQueryService;
+import roomescape.theme.application.service.ThemeCommandService;
+import roomescape.theme.application.service.ThemeQueryService;
 import roomescape.theme.domain.ThemeId;
 import roomescape.theme.ui.dto.CreateThemeWebRequest;
 import roomescape.theme.ui.dto.ThemeResponse;
