+++ src/main/java/roomescape/theme/application/service/ThemeCommandService.java
@@ -1,4 +1,4 @@
-package roomescape.theme.application.usecase;
+package roomescape.theme.application.service;
 
 import roomescape.theme.application.dto.CreateThemeServiceRequest;
 import roomescape.theme.domain.Theme;
