+++ src/main/java/roomescape/time/application/ReservationTimeFacadeImpl.java
@@ -3,8 +3,8 @@ package roomescape.time.application;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import roomescape.time.application.dto.CreateReservationTimeServiceRequest;
-import roomescape.time.application.usecase.ReservationTimeCommandService;
-import roomescape.time.application.usecase.ReservationTimeQueryService;
+import roomescape.time.application.service.ReservationTimeCommandService;
+import roomescape.time.application.service.ReservationTimeQueryService;
 import roomescape.time.domain.ReservationTimeId;
 import roomescape.time.ui.dto.CreateReservationTimeWebRequest;
 import roomescape.time.ui.dto.ReservationTimeResponse;
