+++ src/main/java/roomescape/ui/AdminController.java
@@ -1,4 +1,4 @@
-package roomescape.user.admin.ui;
+package roomescape.ui;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
