+++ src/main/java/roomescape/ui/MemberController.java
@@ -1,4 +1,4 @@
-package roomescape.user.member.ui;
+package roomescape.ui;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -9,5 +9,10 @@ public class MemberController {
     @GetMapping("/reservation")
     public String getReservationPage() {
         return "reservation";
+    }
+
+    @GetMapping("/login")
+    public String getLoginPage() {
+        return "login";
     }
 }
