+++ src/test/java/roomescape/theme/application/ThemeQueryServiceImplTest.java
@@ -10,7 +10,7 @@ import roomescape.reservation.domain.Reservation;
 import roomescape.reservation.domain.ReservationDate;
 import roomescape.reservation.domain.ReservationRepository;
 import roomescape.reservation.domain.ReserverName;
-import roomescape.theme.application.usecase.ThemeQueryServiceImpl;
+import roomescape.theme.application.service.ThemeQueryServiceImpl;
 import roomescape.theme.domain.Theme;
 import roomescape.theme.domain.ThemeDescription;
 import roomescape.theme.domain.ThemeName;
